+++ certbot-ci/certbot_integration_tests/utils/fixtures_function.py
@@ -2,10 +2,69 @@ import os
 import tempfile
 import sys
 import stat
+import subprocess
 
 import pytest
 
 from certbot_integration_tests.utils import misc
+
+
+@pytest.fixture
+def certbot_test_no_force_renew(workspace, config_dir, acme_url, http_01_port, tls_sni_01_port, capsys):
+    def func(args):
+        command = [
+            'certbot', '--server', acme_url, '--no-verify-ssl', '--tls-sni-01-port',
+            str(tls_sni_01_port), '--http-01-port', str(http_01_port),
+            '--manual-public-ip-logging-ok', '--config-dir', config_dir, '--work-dir',
+            os.path.join(workspace, 'work'), '--logs-dir', os.path.join(workspace, 'logs'),
+            '--non-interactive', '--no-redirect', '--agree-tos',
+            '--register-unsafely-without-email', '--debug', '-vv'
+        ]
+
+        command.extend(args)
+
+        # Normally during a test with capture enabled, stdin is set to null
+        # and stdin.isatty returns false.
+        # This non-interactive situation would lead certbot renew operations to wait a random
+        # number of minutes before executing, and we obviously want to avoid that during tests.
+        # So we get real stdin without capture to be used in the subprocess call.
+        with capsys.disabled():
+            real_stdin = sys.stdin
+
+        print('Invoke command:\n{0}'.format(subprocess.list2cmdline(command)))
+        return subprocess.check_output(command, universal_newlines=True, stdin=real_stdin)
+
+    return func
+
+
+@pytest.fixture
+def certbot_test(config_dir, acme_url, http_01_port, tls_sni_01_port):
+    def func(args):
+        command = ['--renew-by-default']
+        command.extend(args)
+        return certbot_test_no_force_renew(command)
+
+    return func
+
+
+@pytest.fixture
+def common_no_force_renew(certbot_test_no_force_renew):
+    def func(args):
+        command = ['--authenticator', 'standalone', '--installer', 'null']
+        command.extend(args)
+        return certbot_test_no_force_renew(command)
+
+    return func
+
+
+@pytest.fixture
+def common(common_no_force_renew):
+    def func(args):
+        command = ['--renew-by-default']
+        command.extend(args)
+        return common_no_force_renew(command)
+
+    return func
 
 
 @pytest.fixture
